refactor: move OpenRouter IPC handler to openrouter-handlers.ts
Extract the OPENROUTER_LIST_MODELS handler from memory-handlers.ts into
its own registerOpenRouterHandlers() function in openrouter-handlers.ts.
Register it alongside other handlers in ipc-handlers/index.ts.
Add OpenRouterModel interface for the parsed API response shape.
