+++ apps/desktop/src/main/ipc-handlers/index.ts
@@ -34,6 +34,7 @@ import { registerProfileHandlers } from './profile-handlers';
 import { registerScreenshotHandlers } from './screenshot-handlers';
 import { registerTerminalWorktreeIpcHandlers } from './terminal';
 import { registerCodexAuthHandlers } from './codex-auth-handlers';
+import { registerOpenRouterHandlers } from './openrouter-handlers';
 import { notificationService } from '../notification-service';
 import { setAgentManagerRef } from './utils';
 
@@ -106,6 +107,9 @@ export function setupIpcHandlers(
   // Memory & infrastructure handlers (for LadybugDB)
   registerMemoryHandlers();
 
+  // OpenRouter model discovery
+  registerOpenRouterHandlers();
+
   // App auto-update handlers
   registerAppUpdateHandlers();
 
@@ -155,5 +159,6 @@ export {
   registerMcpHandlers,
   registerProfileHandlers,
   registerScreenshotHandlers,
-  registerCodexAuthHandlers
+  registerCodexAuthHandlers,
+  registerOpenRouterHandlers
 };
